pyproject.toml
[project]
name = "line-guardrail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=1.9.1",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "aws-lambda-powertools>=2.25.0",
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    "python-dotenv",
]
